Moved popups to app
